Student Score changed to float check

+++ python-basics/reportcardprinter.py
@@ -8,4 +8,4 @@ age = 20
 print(age, type(age))
 
 score = 80.5
-print(isinstance(score, int))
+print(isinstance(score, float))
